refactor(solidlsp): Phase 2 - extract LanguageServerLauncher + dependency providers
Move the LanguageServerDependencyProvider* hierarchy (base, BaseCommand,
SinglePath, Uvx), _create_process_launch_info, _create_language_server_interface,
and initialize-params creation into a new deep module src/solidlsp/ls_launch.py.

Add LanguageServerLauncher collaborator that owns provider creation, launch
info, interface creation, and params building. It calls back into narrow
policy hooks on the facade.

SolidLanguageServer facade:
- Creates the launcher early in __init__
- Delegates launch-related creation to the launcher
- Retains all adapter override hooks unchanged:
  _create_dependency_provider, _create_base_initialize_params,
  _create_language_server_interface, _start_server
- Exposes _dependency_provider as a read-only property shim for
  adapters (e.g. EclipseJDTLS, Matlab) that inspect it after super()

Backward-compat re-exports in solidlsp.ls:
  LanguageServerDependencyProvider* and LanguageServerLauncher so
  ~50 language adapters require zero source changes.

Validation:
- poe lint clean
- poe type-check clean (src/serena src/solidlsp + test)
- poe test -m python: 205 passed (Phase 0 characterization tests preserved)

One shippable commit for Phase 2 per plan e42c071a.
